+++ src/main/java/com/buster/blockbuster/repository/IDao.java
@@ -5,8 +5,8 @@ import java.util.List;
 public interface IDao<T> {
     T save(T t);
     T searchByName(String name);
-    T searchById(Long id);
-    void delete(Long id);
+    T searchById(Integer id);
+    void delete(Integer id);
     List<T> searchAll();
     T update(T t);
 }
